requirements.txt
discogs-client==2.3.0
gspread==5.10.0
google-auth==2.22.0
gspread-formatting==1.1.2
